Configuration file handling (#17)
Added handling of the configuration file.

Closes #5

This adds the field to the action file to look for the configuration
file so we can set a different field, and defaults to
`.github/review.yml`.

Created the basic type for the configuration file.

This adds the:
- rules array
  - no rule types yet
  - they accept an array of include and exclude conditions.
-  the `preventReviewRequests`
- has been renamed from `prevent-review-requests` to
`preventReviewRequests`

Created tests to evaluate all this types and validations.

Also created JOI validations for the configuration object. I have
discovered that it still won't be enough in the case that a regex is
invalid, so I added #16 to have visibility of it.

## Miscelanious

### Created the PullRequestApi class

This class will use the default github secret generated by the action.
We will have a second class to handle the teams.

It is mocked in the tests.

### Created the Runner class

This will be the main class that will combine all the other classes and
APIs (basically the `core.ts` file but with proper abstractions).