Added better rank report and testing (#82)
Enhanced the check report, adding the information of the missing rank.
This resolves #80

Added tests that evaluate the ranking system, and how it applies the
rules. This resolves #81

For the `or` rule, we request the lowest required rank, and for the
`and` rule we request the highest rank.

In the case of `and-distinct` we request the highest rank.