added Pull Request number as input (#94)
This allows is to fetch the PR directly instead of obtaining it from the
event. It is an alternative solution for when we are using a dispatch
event.

This resolves #93

This also allows the use of `environment` because every event is being
triggered from the default branch.

I also updated dependencies.