Refactor of failed reviews objects (#112)
This creates a new type of object: `ReviewFailure`. This object is an
abstract class that contains all the information of the errors. By
having each kind of error implementing it, we can customize the reports.

This was done because reporting problems in the PR is basically the most
exposed part (and the clearer it is, the less that people will ask for
help).

Each class has it's own summary to generate and returns the reviewers
that should be requested. It allows customization per class. It can
solve a lot of problems in the future.

Two error classes were created: `CommonRuleFailure` and
`FellowMissingRankFailure`

### CommonRuleFailure

Wraps all the errors for all the non fellows rules and list what users
are required (and which reviews counted towards fulfilling the rule).

### FellowMissingRankFailure

Quite similar to the `CommonRuleFailure` error, but also handles the
rank required.

In the future I'll add one for `Fellow Score`, which will be a
requirement for detailing the information in #110.

# Other changes

- Renamed structs to contain more clear names.
- Fixed summary not being written for action when check was already
existing
- Simplified `generateCheckRunData` 
- now it is only a couple of lines and the `if` checks are done inside
the errors
- Changed `evaluateCondition` and `andDistinctEvaluation` return type.
- Before they were returning an `[true] | [false, ErrorStruct]` tuple.
After thinking about it, it is simpler to return a `ErrorStruct | null`
type and simply checking if it's null.
- If I would be returning a different object with the `true` condition I
could have keep it, but for this case it was over engineered.
- Changed returned type for `fellowsEvaluation`
    - It now returns the error object directly when applicable.